Refactor thumbnail page retrieval
If there was no thumbnail page defined, then the retrieval would raise a
difficult-to-understand error.  Catch this error and raise a better
error and message.

Move the thumbnail page retrieve to an earlier point.  Where it used to
be, the existing default bitstream(s) for the thumbnail have already
been deleted, which is undesirable if the thumbnail page retrieval
fails.

Closes:  #4